# api/api_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agentics Demo API",
        "version": "0.1.0"
    },
    "paths": {
        "/v1/atypes/": {
            "get": {
                "tags": [
                    "atypes"
                ],
                "summary": "List all saved atypes",
                "description": "Retrieve names of all Pydantic atypes saved in the predefined_types directory.\n\nReturns a simple array of type names (e.g., [\"Answer\", \"Person\"]).",
                "operationId": "list_atypes_v1_atypes__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array",
                                    "title": "Response List Atypes V1 Atypes  Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "atypes"
                ],
                "summary": "Create a new atype",
                "description": "Create a new Pydantic type from either:\n- **field_spec** mode: provide a list of fields with types and metadata\n- **natural_language** mode: describe the type in plain English; an LLM generates the code\n\nOptionally saves the generated .py file to disk (default: save=true).\n\nReturns the generated Python code and JSON schema.",
                "operationId": "create_v1_atypes__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AtypeCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AtypeInfo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/atypes/{name}": {
            "get": {
                "tags": [
                    "atypes"
                ],
                "summary": "Fetch atype details",
                "description": "Retrieve the Python code and JSON schema for a saved atype.\n\nUseful for inspecting existing types before creating an agent.",
                "operationId": "fetch_v1_atypes__name__get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AtypeInfo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "atypes"
                ],
                "summary": "Delete an atype",
                "description": "Remove the .py file for the specified atype from predefined_types.\n\nReturns 204 on success; 404 if the type does not exist.",
                "operationId": "delete_v1_atypes__name__delete",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/atypes/{name}/validate": {
            "post": {
                "tags": [
                    "atypes"
                ],
                "summary": "Validate JSON against atype",
                "description": "Check whether an arbitrary JSON payload conforms to the specified atype schema.\n\nReturns `{\"valid\": true}` on success, or 422 with validation errors.",
                "operationId": "validate_v1_atypes__name__validate_post",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true,
                                "title": "Payload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AtypeCreate": {
                "properties": {
                    "mode": {
                        "$ref": "#/components/schemas/Mode"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "fields": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/FieldSpec"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Fields"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "save": {
                        "type": "boolean",
                        "title": "Save",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "mode",
                    "name"
                ],
                "title": "AtypeCreate"
            },
            "AtypeInfo": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "code": {
                        "type": "string",
                        "title": "Code"
                    },
                    "json_schema": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Json Schema"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "code",
                    "json_schema"
                ],
                "title": "AtypeInfo"
            },
            "FieldSpec": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "type_label": {
                        "type": "string",
                        "title": "Type Label"
                    },
                    "optional": {
                        "type": "boolean",
                        "title": "Optional",
                        "default": true
                    },
                    "use_default": {
                        "type": "boolean",
                        "title": "Use Default",
                        "default": false
                    },
                    "default_value": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Default Value"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "type_label"
                ],
                "title": "FieldSpec"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Mode": {
                "type": "string",
                "enum": [
                    "field_spec",
                    "natural_language"
                ],
                "title": "Mode"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}